requirements.txt
Flask==3.0.0
flask-cors==4.0.0
tensorflow==2.15.0
Pillow==10.1.0
numpy==1.24.3
Werkzeug==3.0.1
groq==0.37.1
